+++ portalsite/portal/models/__init__.py
@@ -1,0 +1,4 @@
+from .measurement import Measurement
+from .model import Model
+from .scoring import Scoring
+from .source import Source
