+++ portalsite/portal/models/model.py
@@ -1,0 +1,48 @@
+"""Data base model: Model"""
+
+# region imports
+# standard
+from typing import TYPE_CHECKING
+
+# 3rd party
+from django.db import models
+from django.urls import reverse
+
+# local
+from portal.core import MODELTYPES
+from .scoring import Scoring
+
+# type hints
+if TYPE_CHECKING:
+    from portal.core.model_type import ModelType
+    from portal.models import Measurement
+# endregion
+
+
+class Model(models.Model):
+    """Prediction model: combines with measurement to create a scoring"""
+
+    name = models.CharField(unique=True, max_length=50)
+    data = models.TextField(help_text='model data (weights, parameters, coefficients, etc.), serialized to string')
+
+    # type interface
+    model_type = models.CharField(max_length=MODELTYPES.id_length, choices=MODELTYPES.choices)
+
+    @property
+    def get_type(self) -> 'ModelType':
+        return MODELTYPES.get(self.model_type)
+
+    def score(self, measurement: 'Measurement') -> 'Scoring':
+        """Returns a new scoring"""
+        scoring = Scoring()
+        scoring.value = self.get_type.score(self, measurement)
+        scoring.model = self
+        scoring.measurement = measurement
+        return scoring
+
+    def __str__(self):
+        return str(self.name)
+
+    def get_absolute_url(self):
+        """Returns the url to display the object."""
+        return reverse('model-detail', args=[str(self.id)])
